+++ grammars/coverage.txt
@@ -1,7 +1,7 @@
 Filename                      Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover
 -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 includes/grammars.h                27                 2    92.59%          10                 1    90.00%          59                16    72.88%
-src/earley/earley.cpp              68                 5    92.65%          10                 1    90.00%         139                14    89.93%
+src/parsing/earley.cpp              68                 5    92.65%          10                 1    90.00%         139                14    89.93%
 src/grammars.cpp                   24                 1    95.83%           1                 0   100.00%          44                 1    97.73%
 tests/earley.cpp                   32                 0   100.00%           3                 0   100.00%         136                 0   100.00%
 tests/grammar.cpp                  15                 0   100.00%          10                 0   100.00%          91                 0   100.00%
