+++ grammars/readme.md
@@ -31,6 +31,23 @@ sudo cp ./*.a /usr/lib
 3. В процессе --- написание алгоритма LR(1).
 4. Предполагалось написать парсеры, выдающие не только заключение о выводимости, но и дерево разбора. За неимением времени это осталось нереализованным, а код был закомментирован, чтобы избежать его учёта в анализе покрытия кода.
 
+#### Структура папки
+##### includes
+* `./grammars.h`: описание класса Grammar и связанных, а также реализация inline-функций
+* `parsing/earley.h`: заголовочный файл для парсера Эрли
+
+#### src
+* `./debug.cpp`: файл, использовавшийся для дебага. Сейчас искючен и _не может быть собран_, так как деревья вывода исключены из проекта.
+* `./grammars.cpp`: общие функции, связанные с грамматиками. Сейчас там только построитель грамматики по описанию.
+* `./main_earley.cpp`: точка входа для standalone парсера Эрли
+* `parsing/earley.cpp`: реализация парсера Эрли
+
+##### includes
+* `./earley.cpp`: тесты алгоритма Эрли
+* `./grammar.cpp`: тесты класса Grammar и общих функций
+* `./main.cpp`: точка входа для тестов
+* `./tests.h`: общий заголовочный файл
+
 #### Описание грамматик
 Описание грамматик должно подчиняться такой структуре:
 ```
@@ -59,7 +76,7 @@ N->***
 ```
 
 ## Тесты
-Написаны тесты с использованием Google Test. Покрытие кода на момент написания алгорита Эрли составило 95.21%. Отчет о покрытии находится в файле ```coverage.txt```.
+Написаны тесты с использованием Google Test. Покрытие кода на момент написания алгорита Эрли составило 95.21%. Отчет о покрытии находится в файле `coverage.txt`.
 
 
 #### Подсчет покрытия
@@ -70,7 +87,7 @@ N->***
 ./count_coverage.sh ./bin/test
 ```
 
-Чтобы посмотреть детальный вывод, необходимо установить флаг ```show```:
+Чтобы посмотреть детальный вывод, необходимо установить флаг `show`:
 ```bash
 ./bin/test
 show=1 ./count_coverage.sh ./bin/test
